refactor: move time duration to common
